Specify type in create() in settings.gradle.kts
The second parameter to create() is a lambda and the type used to get
inferred correctly in IntelliJ Community Edition. But it results in the
following error in IntelliJ Ultimate Edition:

Cannot infer type for type parameter 'S'. Specify it explicitly.

https://youtrack.jetbrains.com/projects/KTIJ/issues/KTIJ-34210 also
suggests that the only workaround at the moment is to specify the type
explicitly.

At the time of this commit, the Community and Ultimate editions were
both the latest available:

IntelliJ IDEA 2025.1.4.1 (Community Edition)
Build #IC-251.27812.49, built on July 24, 2025

IntelliJ IDEA 2025.2.3 (Ultimate Edition)
Build #IU-252.26830.84, built on October 1, 2025